feat(runner+ci): one-command runner, execution tests on 4 live engines, fix 14 scripts
Add bin/dmc-dba: runs a single category or the whole suite read-only against a
live engine and writes a self-contained HTML report (and Markdown) with
per-script pass/fail, impact and timing. Connection via standard env vars or a
full client-command override (works with a containerised client).

Add .github/workflows/test.yml + .github/scripts/start-engine.sh: every script
is now executed against a freshly-started SQL Server 2022, PostgreSQL 16,
MySQL 8 and MongoDB 7 on every push, not just linted.

Fix 14 scripts that passed the linter but failed to run on a clean supported
engine, surfaced by the new execution tests (3 PostgreSQL, 6 MySQL, 5 SQL
Server). All 72 scripts now pass execution across all four engines.

+++ postgresql/storage/table-bloat-estimate.sql
@@ -68,11 +68,11 @@ table_estimates AS (
              * (constants.hdr + cs.data_width
                 + (constants.ma
                    - CASE
-                       WHEN ((constants.hdr + cs.data_width) % constants.ma) = 0
+                       WHEN (((constants.hdr + cs.data_width)::numeric % constants.ma)) = 0
                            THEN constants.ma
-                       ELSE (constants.hdr + cs.data_width) % constants.ma
+                       ELSE ((constants.hdr + cs.data_width)::numeric % constants.ma)
                      END))
-             / (constants.bs - 24))                          AS expected_pages
+             / (constants.bs - 24))::bigint                  AS expected_pages
     FROM pg_class                                            AS c
     JOIN pg_namespace                                        AS n ON n.oid = c.relnamespace
     JOIN columns_stats                                       AS cs
